Fixed hideous mouse click ignore bug.
The bug was that some times when clicking a button in a dialog, the button would depress, but not do anything further. The problem lead back to cd_process_mousetrack() which was added by haibaru to improve buttons so that they act on mouse up, rather than mouse down, and only if the mouse up occurs inside the control. This was a good idea, but for some reason mouse up events were sometimes not being detected, so the button would remain stuck in the depressed state without really firing. 

Now, we use brute force to find out where the mouse is if we can't get a mouse up event. It's not ideal, but seems to work far more robustly.